Split RecordDispatcher in interface and implementation class Rename OffsetManager in RecordDispatcherListener
