+++ data-plane/dispatcher/src/main/java/dev/knative/eventing/kafka/broker/dispatcher/consumer/impl/UnorderedOffsetManager.java
@@ -15,7 +15,7 @@
  */
 package dev.knative.eventing.kafka.broker.dispatcher.consumer.impl;
 
-import dev.knative.eventing.kafka.broker.dispatcher.consumer.OffsetManager;
+import dev.knative.eventing.kafka.broker.dispatcher.RecordDispatcherListener;
 import io.vertx.core.Future;
 import io.vertx.kafka.client.common.TopicPartition;
 import io.vertx.kafka.client.consumer.KafkaConsumer;
@@ -32,7 +32,7 @@ import org.slf4j.LoggerFactory;
 /**
  * This class implements the offset strategy that makes sure that, even unordered, the offset commit is ordered.
  */
-public final class UnorderedOffsetManager implements OffsetManager {
+public final class UnorderedOffsetManager implements RecordDispatcherListener {
 
   private static final Logger logger = LoggerFactory
     .getLogger(UnorderedOffsetManager.class);
